refactor(core): Rename files with kebab-case convention
