+++ core/ports/document-repository.ts
